State vars reduction

+++ contracts/test/ReentrancyAttacker.sol
@@ -17,7 +17,7 @@ interface ITokenomics {
     /// @dev Deposits ETH from protocol-owned services in batch.
     /// @param serviceIds Set of service Ids.
     /// @param amounts Set of corresponding amounts deposited on behalf of each service Id.
-    function depositETHFromServices(uint256[] memory serviceIds, uint256[] memory amounts) external payable;
+    function depositETHFromServices(uint32[] memory serviceIds, uint96[] memory amounts) external payable;
 }
 
 contract ReentrancyAttacker {
@@ -79,7 +79,7 @@ contract ReentrancyAttacker {
     }
 
     /// @dev Lets the attacker call back its contract to get back to the depositETHFromServices() function.
-    function badDepositETHFromServices(uint256[] memory serviceIds, uint256[] memory amounts) external payable
+    function badDepositETHFromServices(uint32[] memory serviceIds, uint96[] memory amounts) external payable
     {
         attackOnDepositETHFromServices = true;
         ITokenomics(treasury).depositETHFromServices{value: msg.value}(serviceIds, amounts);
